Reference

Bonobo.AbstractTraverseStrategy — Type

AbstractTraverseStrategy

The abstract type for a traverse strategy. If you implement a new traverse strategy this must be the supertype.

If you want to implement your own strategy the get_next_node function needs a new method which dispatches on the traverse_strategy argument.

Bonobo.BFS — Type

BFS <: AbstractTraverseStrategy

The BFS traverse strategy always picks the node with the lowest bound first. If there is a tie then the smallest node id is used as a tie breaker.

Bonobo.add_new_solution! — Method

add_new_solution!(tree::BnBTree{N,R,V,S}, node::AbstractNode) where {N,R,V,S<:DefaultSolution{N,V}}

Currently it changes the general solution itself by calling get_relaxed_values which needs to be implemented by you.

Todo: Add a possibility to store several solutions based on some options.

This function needs to be implemented by you if you have a different type of Solution object than DefaultSolution.

Bonobo.add_node! — Method

add_node!(tree::BnBTree{Node}, node_info::NamedTuple)

Add a new node to the tree using the node_info. For information on that see set_root!.

Bonobo.bound! — Method

bound!(tree::BnBTree, current_node_id)

Close all nodes which have a lower bound higher or equal to the incumbent

Bonobo.close_node! — Method

close_node!(tree::BnBTree, node::AbstractNode)

Delete the node from the priority queue from the nodes

Bonobo.create_node — Method

create_node(Node, node_id::Int, node_info::NamedTuple)

Creates a node of type Node with id node_id and the named tuple node_info. For information on that see set_root!.

Bonobo.get_next_node — Method

get_next_node(tree::BnBTree, travese_strategy::BFS)

Get the next node of the tree which shall be evaluted next by evaluate_node. If you want to implement your own traversing strategy check out AbstractTraverseStrategy.

Bonobo.get_objective_value — Method

get_objective_value(tree::BnBTree; result=1)

Return the objective value

Bonobo.get_relaxed_values — Method

get_relaxed_values(tree::BnBTree, node::AbstractNode)

Get the values of the current node. This is always called only after evaluate_node is called. It is used to store a Solution object. Return the type of Value given to the initialize method.

Bonobo.get_solution — Method

get_solution(tree::BnBTree; result=1)

Return the solution values of the problem. See get_objective_value to obtain the objective value instead.

Bonobo.initialize — Method

initialize(; kwargs...)

Initialize the branch and bound framework with the the following arguments. Later it can be dispatched on BnBTree{Node, Root, Solution} for various methods.

Keyword arguments

traverse_strategy [BFS] currently the only supported traverse strategy is BFS. Should be an AbstractTraverseStrategy

Node DefaultNode can be special structure which is used to store all information about a node.

needs to have AbstractNode as the super type

needs to have std :: BnBNode as a field (see BnBNode)

Solution DefaultSolution stores the node and several other information about a solution

root [nothing] the information about the root problem. The type can be used for dispatching on types

sense [:Min] can be :Min or :Max depending on the objective sense

Value [Vector{Float64}] the type of a solution

Return a BnBTree object which is the input for optimize!.

Bonobo.optimize! — Method

optimize!(tree::BnBTree)

Optimize the problem using a branch and bound approach.

The steps are the following:

while !terminated(tree) # as long as there are open nodes
    # get the next open node depending on the traverse strategy
    node = get_next_node(tree, tree.options.traverse_strategy) 
    # needs to be implemented by you
    # Should evaluate the current node and return the lower and upper bound
    # if the problem is infeasible both values should be set to NaN
    lb, ub = evaluate_node!(tree, node) 
    # updates the upper and lower bound of the node struct
    set_node_bound!(tree.sense, node, lb, ub)

    # update the best solution 
    updated = update_best_solution!(tree, node)
    updated && bound!(tree, node.id)
    
    # remove the current node
    close_node!(tree, node)
    # needs to be implemented by you
    # create branches from the current node
    branch!(tree, node)
end

every function of the above can be overriden by your own method.

Bonobo.set_node_bound! — Method

set_node_bound!(objective_sense::Symbol, node::AbstractNode, lb, ub)

Set the bounds of the node object to the lower and upper bound given. Internally everything is stored as a minimization problem. Therefore the objective_sense :Min/:Max is needed.

Bonobo.set_root! — Method

set_root!(tree::BnBTree, node_info::NamedTuple)

Set the root node information based on the node_info which needs to include the same fields as the Node struct given to the initialize method. (Besides the std field which is set by Bonobo automatically)

Example

If your node structure is the following:

mutable struct MIPNode <: AbstractNode
    std :: BnBNode
    lbs :: Vector{Float64}
    ubs :: Vector{Float64}
    status :: MOI.TerminationStatusCode
end

then you can call the function with this syntax:

BB.set_root!(bnb_model, (
    lbs = fill(-Inf, length(x)),
    ubs = fill(Inf, length(x)),
    status = MOI.OPTIMIZE_NOT_CALLED
))

Bonobo.terminated — Method

terminated(tree::BnBTree)

Return true when the branch and bound loop in optimize! should be terminated. Default behavior is to terminate the loop only when no nodes exist in the priority queue.

Bonobo.update_best_solution! — Method

update_best_solution!(tree::BnBTree, node::AbstractNode)

Update the best solution when we found a better incumbent. Calls [add_new_solution!] if this is the case