Reference

Bonobo.initialize — Method

initialize(; kwargs...)

Initialize the branch and bound framework with the the following arguments. Later it can be dispatched on BnBTree{Node, Root, Solution} for various methods.

Keyword arguments

traverse [:BFS] currently the only supported traverse strategy is BFS.

Node DefaultNode can be special structure which is used to store all information about a node.

needs to have AbstractNode as the super type

needs to have std :: BnBNode as a field (see BnBNode)

Solution DefaultSolution stores the node and several other information about a solution

root [nothing] the information about the root problem. The type can be used for dispatching on types

sense [:Min] can be :Min or :Max depending on the objective sense

Value [Vector{Float64}] the type of a solution

Return a BnBTree object which is the input for optimize!.

Bonobo.optimize! — Method

optimize!(tree::BnBTree)

Optimize the problem using a branch and bound approach.

The steps are the following:

while !terminated(tree) # as long as there are open nodes
    # get the next open node depending on the traverse strategy
    node = get_next_node(tree) 
    # needs to be implemented by you
    # Should evaluate the current node and return the lower and upper bound
    # if the problem is infeasible both values should be set to NaN
    lb, ub = evaluate_node!(tree, node) 
    # updates the upper and lower bound of the node struct
    set_node_bound!(tree.sense, node, lb, ub)

    # update the best bound of the tree depending on the new upper bound of the node
    # calls bound! to remove nodes from the tree with a worse bound
    update_best_bound!(tree, node)
    # update the best solution 
    update_best_solution!(tree, node)
    
    # remove the current node
    close_node!(tree, node)
    # needs to be implemented by you
    # create branches from the current node
    branch!(tree, node)
end

every function of the above can be overriden by your own method.